Small update for the line for smallest pie sector
Small update for the line for the smallest Pie Chart sector -
increasing it further to look make it better, still no luck for the
labels for the Darts Chart. I can make them have enough line spacing
only if I mix up the label order, which is not what I want to do.